src/requirements.txt
Django
djangorestframework
markdown
django-filter
djangorestframework_simplejwt
django-redis
PyJWT
psycopg2
gunicorn